ughhhhh

+++ VoxelTest2/Primitives/Geometry.cs
@@ -28,20 +28,21 @@ namespace VoxelTest2.Primitives
             new Vector3( 0.5f,  0.5f,  0.5f), // Top-right-front
             new Vector3(-0.5f,  0.5f,  0.5f)  // Top-left-front
         };
-        public int[] blockIndices = new int[]
-        {
-            // Back face
-            0, 1, 2, 2, 3, 0,
-            // Front face
-            4, 5, 6, 6, 7, 4,
-            // Left face
-            0, 3, 7, 7, 4, 0,
-            // Right face
-            1, 2, 6, 6, 5, 1,
-            // Bottom face
-            0, 1, 5, 5, 4, 0,
-            // Top face
-            2, 3, 7, 7, 6, 2
-        };
+            public float[] blockIndices = new float[]
+            {
+                // Back face
+                0, 1, 2, 2, 3, 0,
+                // Front face
+                4, 5, 6, 6, 7, 4,
+                // Left face
+                0, 3, 7, 7, 4, 0,
+                // Right face
+                1, 2, 6, 6, 5, 1,
+                // Bottom face
+                0, 1, 5, 5, 4, 0,
+                // Top face
+                2, 3, 7, 7, 6, 2
+            };
     }
 }
+
